Add some helper functions for mocking methods and signals in tests
