+++ crates/cxx-qt-gen/src/parser/signals.rs
@@ -14,6 +14,9 @@ use proc_macro2::TokenStream;
 use quote::quote;
 use syn::{spanned::Spanned, Attribute, Error, ForeignItemFn, Ident, Result, Visibility};
 
+#[cfg(test)]
+use quote::format_ident;
+
 #[derive(Clone)]
 /// Describes an individual Signal
 pub struct ParsedSignal {
@@ -55,6 +58,23 @@ impl ParsedSignal {
             inherit: false,
             private: false,
             docs,
+        }
+    }
+
+    #[cfg(test)]
+    /// Test fn for creating a dummy signal from a method body
+    pub fn from_method(method: &ForeignItemFn) -> Self {
+        Self {
+            method: method.clone(),
+            qobject_ident: format_ident!("MyObject"),
+            mutable: true,
+            parameters: vec![],
+            name: Name::from_rust_ident_and_attrs(&method.sig.ident, &method.attrs, None, None)
+                .unwrap(),
+            safe: true,
+            inherit: false,
+            private: false,
+            docs: vec![],
         }
     }
